Moved API out of an unnecessary folder
