+++ profiling/examples/profiles.rs
@@ -1,8 +1,8 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
 
+use datadog_profiling::api;
 use datadog_profiling::internal::Profile;
-use datadog_profiling::profile::api;
 use std::io::Write;
 use std::process::exit;
 use std::time::SystemTime;
