+++ profiling/src/profile/mod.rs
@@ -1,7 +1,6 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
 
-pub mod api;
 pub mod profiled_endpoints;
 
 pub type Timestamp = std::num::NonZeroI64;
