+++ profiling/tests/wordpress.rs
@@ -1,6 +1,7 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
 
+use datadog_profiling::api;
 use datadog_profiling::pprof;
 use datadog_profiling::profile::*;
 use lz4_flex::frame::FrameDecoder;
